fix import issues by introducing a ci.common module
