+++ lib/spack/spack/ci/formatters/gitlab.py
@@ -15,7 +15,7 @@ import spack.binary_distribution as bindist
 import spack.config as cfg
 import spack.util.spack_yaml as syaml
 
-from .. import (
+from ..common import (
     SPACK_RESERVED_TAGS,
     PipelineDag,
     PipelineOptions,
